+++ omero/sysadmins/unix/walkthrough/setup_omero_nginx.sh
@@ -1,9 +1,0 @@
-#!/bin/bash
-NGINXCMD=${1:-nginx}
-set -e -u -x
-
-source `dirname $0`/settings-web.env
-
-#start-config
-OMERO.server/bin/omero config set omero.web.application_server wsgi-tcp
-OMERO.server/bin/omero web config $NGINXCMD --http "$WEBPORT" > OMERO.server/nginx.conf.tmp