+++ omero/sysadmins/unix/walkthrough/setup_omero_nginx50.sh
@@ -1,8 +1,0 @@
-#!/bin/bash
-set -e -u -x
-
-source `dirname $0`/settings-web.env
-
-#start-config
-OMERO.server/bin/omero config set omero.web.application_server fastcgi-tcp
-OMERO.server/bin/omero web config nginx --system --http "$WEBPORT" > OMERO.server/nginx.conf.tmp