+++ omero/sysadmins/unix/walkthrough/walkthrough_centos7.sh
@@ -1,146 +1,104 @@
 #!/bin/bash
 set -e -u -x
-source settings.env
-source settings-web.env
 
 #start-step01: As root, install dependencies
-
 yum -y install epel-release
 
-# installed for convenience
 yum -y install unzip wget bc
 
 # install Java
-yum -y install java-1.8.0-openjdk
+yum -y install java-11-openjdk
 
 # install dependencies
 
-yum -y install python-{pip,devel,virtualenv,yaml,jinja2,tables}
-
-
-#start-web-dependencies
-yum -y install python-pillow numpy
-#end-web-dependencies
+yum -y install python3
+#end-step01
 # install Ice
 #start-recommended-ice
 curl -sL https://zeroc.com/download/Ice/3.6/el7/zeroc-ice3.6.repo > \
 /etc/yum.repos.d/zeroc-ice3.6.repo
-yum -y install gcc-c++
-yum -y install libdb-utils
-yum -y install openssl-devel bzip2-devel
 
 yum -y install ice-all-runtime
-
-pip install "zeroc-ice>3.5,<3.7"
 #end-recommended-ice
 
 
 # install Postgres
-# Postgres, reconfigure to allow TCP connections
-yum -y install http://yum.postgresql.org/9.6/redhat/rhel-7-x86_64/pgdg-redhat96-9.6-3.noarch.rpm
-yum -y install postgresql96-server postgresql96
+yum -y install https://yum.postgresql.org/11/redhat/rhel-7-x86_64/pgdg-redhat11-11-2.noarch.rpm
+yum -y install postgresql11-server postgresql11
 
-PGSETUP_INITDB_OPTIONS=--encoding=UTF8 /usr/pgsql-9.6/bin/postgresql96-setup initdb
+PGSETUP_INITDB_OPTIONS=--encoding=UTF8 /usr/pgsql-11/bin/postgresql-11-setup initdb
 
-sed -i.bak -re 's/^(host.*)ident/\1md5/' /var/lib/pgsql/9.6/data/pg_hba.conf
-systemctl start postgresql-9.6.service
+sed -i.bak -re 's/^(host.*)ident/\1md5/' /var/lib/pgsql/11/data/pg_hba.conf
+systemctl start postgresql-11.service
 
-systemctl enable postgresql-9.6.service
-
+systemctl enable postgresql-11.service
 #end-step01
 
-#start-step02: As root, create an omero system user and directory for the OMERO repository
-useradd -m omero
+#start-step02: As root, create a local omero-server system user and directory for the OMERO repository
+useradd -mr omero-server
 # Give a password to the omero user
-# e.g. passwd omero
-chmod a+X ~omero
+# e.g. passwd omero-server
+chmod a+X ~omero-server
 
 mkdir -p "$OMERO_DATA_DIR"
-chown omero "$OMERO_DATA_DIR"
+chown omero-server "$OMERO_DATA_DIR"
 #end-step02
-
 #start-step03: As root, create a database user and a database
-
 echo "CREATE USER $OMERO_DB_USER PASSWORD '$OMERO_DB_PASS'" | su - postgres -c psql
 su - postgres -c "createdb -E UTF8 -O '$OMERO_DB_USER' '$OMERO_DB_NAME'"
 
 psql -P pager=off -h localhost -U "$OMERO_DB_USER" -l
 #end-step03
 
-#start-step04: As the omero system user, install the OMERO.server
-#start-copy-omeroscript
-cp settings.env settings-web.env step04_all_omero.sh setup_omero_db.sh ~omero 
-#end-copy-omeroscript
+#start-step03bis: As root, create a virtual env and install dependencies
+# Create a virtual env and activate it
+python3 -mvenv $VENV_SERVER
+
+# Install the Ice Python binding
+$VENV_SERVER/bin/pip install https://github.com/ome/zeroc-ice-py-centos7/releases/download/0.2.1/zeroc_ice-3.6.5-cp36-cp36m-linux_x86_64.whl
+#end-step03bis
+
+#start-step04-pre: As root, install omero-py and download the OMERO.server
+# Install omero-py
+$VENV_SERVER/bin/pip install "omero-py>=5.6.0"
 #start-release-ice36
-cd ~omero
-SERVER=https://downloads.openmicroscopy.org/latest/omero5/server-ice36.zip
-wget $SERVER -O OMERO.server-ice36.zip
+cd /opt/omero/server
+SERVER=https://downloads.openmicroscopy.org/omero/5.6/server-ice36.zip
+wget -q $SERVER -O OMERO.server-ice36.zip
 unzip -q OMERO.server*
 #end-release-ice36
+# change ownership of the folder
+chown -R omero-server OMERO.server-*
 ln -s OMERO.server-*/ OMERO.server
-OMERO.server/bin/omero config set omero.data.dir "$OMERO_DATA_DIR"
-OMERO.server/bin/omero config set omero.db.name "$OMERO_DB_NAME"
-OMERO.server/bin/omero config set omero.db.user "$OMERO_DB_USER"
-OMERO.server/bin/omero config set omero.db.pass "$OMERO_DB_PASS"
-OMERO.server/bin/omero db script -f OMERO.server/db.sql --password "$OMERO_ROOT_PASS"
-psql -h localhost -U "$OMERO_DB_USER" "$OMERO_DB_NAME" < OMERO.server/db.sql
+#end-step04-pre
+
+#start-step04: As the omero-server system user, configure it
+#start-copy-omeroscript
+cp settings.env step04_all_omero.sh setup_omero_db.sh ~omero 
+#end-copy-omeroscript
+omero config set omero.data.dir "$OMERO_DATA_DIR"
+omero config set omero.db.name "$OMERO_DB_NAME"
+omero config set omero.db.user "$OMERO_DB_USER"
+omero config set omero.db.pass "$OMERO_DB_PASS"
+omero db script -f $OMERODIR/db.sql --password "$OMERO_ROOT_PASS"
+psql -h localhost -U "$OMERO_DB_USER" "$OMERO_DB_NAME" < $OMERODIR/db.sql
 #end-step04
 
-#start-step05: As omero, install OMERO.web dependencies
-#web-requirements-recommended-start
-pip install -r OMERO.server/share/web/requirements-py27.txt
-#web-requirements-recommended-end
-#start-configure-nginx: As the omero system user, configure OMERO.web
-OMERO.server/bin/omero config set omero.web.application_server wsgi-tcp
-OMERO.server/bin/omero web config nginx --http "$WEBPORT" > OMERO.server/nginx.conf.tmp
-#end-configure-nginx
-# As root, install nginx
-#start-nginx-install
-yum -y install nginx
-#end-nginx-install
-#start-nginx-admin
-sed -i.bak -re 's/( default_server.*)/; #\1/' /etc/nginx/nginx.conf
 
-cp OMERO.server/nginx.conf.tmp /etc/nginx/conf.d/omero-web.conf
-
-systemctl enable nginx
-
-systemctl start nginx
-
-#end-nginx-admin
-
-#end-step05
-
-#start-step06: As root, run the scripts to start OMERO and OMERO.web automatically
-cp omero-systemd.service /etc/systemd/system/omero.service
-cp omero-web-systemd.service /etc/systemd/system/omero-web.service
+#start-step06: As root, run the scripts to start OMERO automatically
+cp omero-server-systemd.service /etc/systemd/system/omero-server.service
 
 systemctl daemon-reload
 
-systemctl enable omero.service
-systemctl enable omero-web.service
+systemctl enable omero-server.service
 #end-step06
 
 #start-step07: As root, secure OMERO
-chmod go-rwx OMERO.server/etc OMERO.server/var
+chmod go-rwx $OMERODIR/etc $OMERODIR/var
 
 # Optionally restrict access to the OMERO data directory
 # chmod go-rwx "$OMERO_DATA_DIR"
 #end-step07
-
-#start-step08: As root, perform regular tasks
-#start-omeroweb-cron
-OMERO_USER=omero
-OMERO_SERVER=/home/omero/OMERO.server
-su - ${OMERO_USER} -c "${OMERO_SERVER}/bin/omero web clearsessions"
-#end-omeroweb-cron
-#Copy omero-web-cron into the appropriate location
-#start-copy-omeroweb-cron
-
-cp omero-web-cron /etc/cron.daily/omero-web
-chmod a+x /etc/cron.daily/omero-web
-#end-copy-omeroweb-cron
-#end-step08
 #start-selinux
 
 if [ $(getenforce) != Disabled ]; then