+++ omero/changelog.rst
@@ -5,12 +5,12 @@ CHANGELOGS
 Links to decoupled repositories
 ===============================
 
-Since OMERO 5.5, the following repositories have been decoupled.
+Starting from OMERO 5.5, the following repositories have been decoupled.
 
 omero-build:
 ------------
 
-- omero-java-gateway :java_gateway_source:`CHANGELOG.md<CHANGELOG.md>`
+- omero-gateway-java :java_gateway_source:`CHANGELOG.md<CHANGELOG.md>`
 - omero-blitz :blitz_source:`CHANGELOG.md<CHANGELOG.md>`
 - omero-server :server_source:`CHANGELOG.md<CHANGELOG.md>`
 - omero-renderer :renderer_source:`CHANGELOG.md<CHANGELOG.md>`
@@ -22,4 +22,5 @@ omero clients:
 --------------
 - omero-matlab :matlab_source:`CHANGELOG.md<CHANGELOG.md>`
 - omero-insight :insight_source:`CHANGELOG.md<CHANGELOG.md>`
-
+- omero-py :py_source:`CHANGELOG.md<CHANGELOG.md>`
+- omero-web :web_source:`CHANGELOG.md<CHANGELOG.md>`